+++ python/cudf_polars/tests/expressions/test_booleanfunction.py
@@ -12,8 +12,8 @@ from cudf_polars.testing.asserts import (
     assert_gpu_result_equal,
     assert_ir_translation_raises,
 )
+from cudf_polars.testing.engine_utils import is_streaming_engine
 from cudf_polars.utils.versions import POLARS_VERSION_LT_132
-from tests.conftest import is_streaming_engine
 
 if TYPE_CHECKING:
     from collections.abc import Callable
