+++ python/cudf_polars/tests/test_dataframescan.py
@@ -13,8 +13,8 @@ from cudf_polars.testing.asserts import (
     assert_gpu_result_equal,
     assert_ir_translation_raises,
 )
+from cudf_polars.testing.engine_utils import is_streaming_engine
 from cudf_polars.utils.versions import POLARS_VERSION_LT_138
-from tests.conftest import is_streaming_engine
 
 
 @pytest.mark.parametrize(
