+++ python/cudf_polars/tests/test_groupby.py
@@ -15,13 +15,13 @@ from cudf_polars.testing.asserts import (
     assert_gpu_result_equal,
     assert_ir_translation_raises,
 )
+from cudf_polars.testing.engine_utils import get_blocksize_mode, is_streaming_engine
 from cudf_polars.utils.versions import (
     POLARS_VERSION_LT_132,
     POLARS_VERSION_LT_134,
     POLARS_VERSION_LT_136,
     POLARS_VERSION_LT_1321,
 )
-from tests.conftest import get_blocksize_mode, is_streaming_engine
 
 
 @pytest.fixture
